Extract `getAllowedPluginActions` to `utils` folder
